+++ Project_wsl_2026/Software/main_tennsy4.0/lib/Motors/serial/DSR1202.hpp
@@ -5,14 +5,14 @@
 #include "Motors.hpp"
 
 // モータ用のトグルスイッチのピン番号を設定
-void DSR1202_set_motormove_togglepin(int pin);
+void DSR1202_set_motormove_togglepin(uint8_t pin);
 
 // シリアル定義・開始
-void DSR1202_init(HardwareSerial *serial, int baudrate);
+void DSR1202_init(HardwareSerial *serial, uint32_t baudrate);
 
 // すべてのモーターを停止させる
 void DSR1202_break();
 // それぞれのチャンネルを動かす
-void DSR1202_move(int value_1ch, int value_2ch, int value_3ch, int value_4ch);
+void DSR1202_move(uint8_t value_1ch, uint8_t value_2ch, uint8_t value_3ch, uint8_t value_4ch);
 
 #endif
